backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
groq==1.1.0
ffmpeg-python==0.2.0
pymongo==4.7.2
rank-bm25==0.2.2
wikipedia-api==0.6.0
google-generativeai==0.7.2
tensorflow==2.15.0
librosa==0.10.1
tf-keras==2.15.1
opencv-python==4.9.0.80
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4
gunicorn